+++ src/key_stat_/index.js
@@ -1,0 +1,22 @@
+import { stock_stats__fetch_get } from '../stock_stats/index.js'
+/** @typedef {import('iex-cloud').KeyStat}KeyStat */
+/** @typedef {import('../_types').iex_fetch__params_T}iex_fetch__params_T */
+/**
+ * @param {string}ticker
+ * @param {iex_fetch__params_T}[iex_fetch__params]
+ * @returns {Promise<KeyStat|null>}
+ * @private
+ */
+export async function key_stat_(
+	ticker,
+	iex_fetch__params = {}
+) {
+	const [
+		payload,
+		res
+	] = await stock_stats__fetch_get(ticker, iex_fetch__params)
+	return res.ok ? payload : null
+}
+export {
+	key_stat_ as stock_stats_
+}