+++ src/stock_stats/index.js
@@ -1,5 +1,6 @@
 import { iex__fetch_get } from '../iex__fetch_get/index.js'
 /** @typedef {import('@ctx-core/error').HttpError}HttpError */
+/** @typedef {import('iex-cloud').KeyStat}KeyStat */
 /**
  * @param {string}ticker
  * @returns {string}
@@ -11,7 +12,7 @@ export function stock_stats_path_(ticker) {
 /**
  * @param {string}ticker
  * @param {import('../_types').iex_fetch__params_T}[iex_fetch__params]
- * @returns {Promise<[{ marketcap:string }, Response]|[HttpError, Response]>}
+ * @returns {Promise<[KeyStat, Response]|[HttpError, Response]>}
  * @private
  */
 export function stock_stats__fetch_get(
@@ -19,20 +20,4 @@ export function stock_stats__fetch_get(
 	iex_fetch__params = {}
 ) {
 	return iex__fetch_get(stock_stats_path_(ticker), iex_fetch__params)
-}
-/**
- * @param {string}ticker
- * @param {import('../_types').iex_fetch__params_T}[iex_fetch__params]
- * @returns {Promise<string|null>}
- * @private
- */
-export async function stock_stats_(
-	ticker,
-	iex_fetch__params = {}
-) {
-	const [
-		payload,
-		res
-	] = await stock_stats__fetch_get(ticker, iex_fetch__params)
-	return res.ok ? payload.marketcap : null
 }